Use a CharSequence as a source for read/writeString
This patch allows the usage of the interface CharSequence
as a source of output as read from / written to byte arrays,
which allows the usage of StringBuilders for example.
This avoids the overhead of allocations of strings when especially when there
are millions of objects being read/ written per second.