Avoid interception of generic Exception
Overall, restore a good practice of catching the desired exception types
explicitly, in order to avoid accidental interception of exceptions that
can be meaningful to API caller.

In particular, do not hide subtypes of IOException which are good
indicators of broken data. See GitHub issue #291 for more details.

In order to simplify the enforcement of this good practice - use
Accessor implementations wherever possible.